Factor out shared.game_utils
